+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - [QuickFormInterface::getId() is replaced by QuickFormInterface::getQuickId()](https://www.drupal.org/node/3379686)
 - [Remove deprecated farm_log_asset_names_summary()](https://www.drupal.org/node/3359456)
 - [Issue #3394069: Update quantities to use bundle permission granularity](https://www.drupal.org/node/3394069)
+- [Issue #3357679: Allow material quantities to reference multiple material types](https://www.drupal.org/project/farm/issues/3357679)
 
 ## [2.2.2] 2023-10-25