+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - [Remove deprecated farm_log_asset_names_summary()](https://www.drupal.org/node/3359456)
 - [Issue #3394069: Update quantities to use bundle permission granularity](https://www.drupal.org/node/3394069)
 - [Issue #3357679: Allow material quantities to reference multiple material types](https://www.drupal.org/project/farm/issues/3357679)
+- [Issue #3330490: Update Drupal core to 10.x in farmOS](https://www.drupal.org/project/farm/issues/3330490)
 
 ## [2.2.2] 2023-10-25